add default structure
Code directory structure is inside `src` directory.
Namespaces are mapped into it as follows:
- `Deliberry\${ProjectNamespace}` -> `src/Domain`.
   Domain and business-related code.
- `Deliberry\Application` `src/Application`: glue code between entrypoints and domain.
- `Deliberry\Infrastructure` -> `src/Infrastructure`: connect with external services, like databases, apis, etc.
- `Deliberry\Bundle` -> `src/Infrastructure/Framework`: Framework related code goes here, like bundles and controllers.

All code except the domain is specific to this application and must not be shared between projects.
If there's a need to share code between projects is better to extract to a new library/bundle and add a dependency to it.